refactor: update settings
